avatars/requirements.txt
aiobotocore==2.23.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aioitertools==0.12.0
aiosignal==1.4.0
annotated-types==0.7.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.9.0
attrs==25.3.0
boto3==1.38.0
botocore==1.38.27
certifi==2025.6.15
cffi==1.17.1
cfgv==3.4.0
distlib==0.3.9
Django==5.2.4
django-storages==1.14.6
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.5.1
identify==2.6.12
idna==3.10
jmespath==1.0.1
minio==7.2.15
multidict==6.6.3
nodeenv==1.9.1
pillow==11.3.0
platformdirs==4.3.8
pre_commit==4.2.0
propcache==0.3.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
s3fs==2025.5.1
s3transfer==0.12.0
six==1.17.0
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.31.2
wrapt==1.17.2
yarl==1.20.1
